Add a common fix for isBuiltin method usage
