+++ lib/internal/Magento/Framework/Code/Reader/ClassReader.php
@@ -5,6 +5,7 @@
  */
 namespace Magento\Framework\Code\Reader;
 
+use Magento\Framework\GetParameterClassTrait;
 use ReflectionClass;
 use ReflectionException;
 use ReflectionParameter;
@@ -14,6 +15,11 @@ use ReflectionParameter;
  */
 class ClassReader implements ClassReaderInterface
 {
+    use GetParameterClassTrait;
+
+    /**
+     * @var array
+     */
     private $parentsCache = [];
 
     /**
@@ -54,26 +60,6 @@ class ClassReader implements ClassReaderInterface
         }
 
         return $result;
-    }
-
-    /**
-     * Get class by reflection parameter
-     *
-     * @param ReflectionParameter $reflectionParameter
-     * @return ReflectionClass|null
-     * @throws ReflectionException
-     */
-    private function getParameterClass(ReflectionParameter $reflectionParameter): ?ReflectionClass
-    {
-        $parameterType = $reflectionParameter->getType();
-        // In PHP8, $parameterType could be an instance of ReflectionUnionType, which doesn't have isBuiltin method.
-        if ($parameterType !== null && method_exists($parameterType, 'isBuiltin') === false) {
-            return null;
-        }
-
-        return $parameterType && !$parameterType->isBuiltin()
-            ? new ReflectionClass($parameterType->getName())
-            : null;
     }
 
     /**
